requirements.txt
asgiref==3.2.7
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
client==0.0.1
Django==3.0.4
gcloud==0.18.3
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-core==1.4.3
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
jws==0.1.3
midtransclient==1.1.1
mysql==0.0.2
mysqlclient==1.4.6
oauth2client==4.1.2
oauthlib==3.1.0
Pillow==7.2.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.9
Pyrebase4==4.4.1
PySocks==1.7.1
python-jwt==2.0.1
pytz==2019.3
requests==2.25.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
six==1.15.0
sqlparse==0.3.1
tweepy==3.9.0
urllib3==1.25.10
